requirements.txt
numpy 
boto3 
s3fs
botocore 
openpyxl
geopandas